fix : s3 업로드 가능 형태로 프론트 수정
